Simulate microbiome perturbation data

To simulate microbiome longitudinal count data for perturbation experiments with user defined starting and ending time point, effect size and duration of perturbation

sim_ts_perturb(
  endtime,
  n_species,
  n_depth,
  Sigma = NULL,
  beta,
  t1,
  t2,
  time,
  species,
  duration
)

Arguments

endtime

the number of time points of the sample

n_species

number of species

n_depth

sequence depth

Sigma

correlation matrix for muti-nomal distribution

beta

effect of perturbation

t1

time of reaching the maximum of perturbation

t2

time of ending the maximum of perturbation

time

the start time for perturbation

species

perturbed species

duration

the duration of perturbation

Value

the simulated data

Examples

sim_ts_perturb(endtime=30, n_species=30, 1000, Sigma=NULL, beta=1,species=c(1,2,3),time=4, t1=8,t2=12,duration = 16)
#>     V1  V2  V3 V4 V5 V6 V7 V8 V9 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V20
#> 1   23  30  31 66 20 36 86 28 78  40  26  16  37  16  24  19  28  19  29  41
#> 2   37  33  41 48 22 40 48 30 55  36  24  27  41  24  16  22  41  23  34  39
#> 3   38  43  22 32 24 35 44 34 41  52  31  34  40  26  35  34  31  31  36  39
#> 4   47  36  30 40 36 47 44 29 40  41  30  34  38  29  29  28  35  45  38  36
#> 5   82  79  76 33 28 34 22 31 30  23  22  34  24  30  34  18  36  29  36  25
#> 6  190 192 191 22 12  9 13 15  8  12   6  15  18  24  17  17  16  17  27  16
#> 7  286 313 274  3  4  3  6  3  4   8   6   5   3   3   7   5   5   7   4   6
#> 8  336 317 322  1  1  1  0  2  0   0   2   0   0   3   0   1   0   2   1   1
#> 9  369 312 310  0  1  0  1  0  2   1   0   0   0   0   0   0   0   0   0   1
#> 10 327 346 323  0  0  0  1  1  0   0   0   0   0   0   0   0   0   0   0   0
#> 11 317 359 321  0  0  0  0  0  0   0   0   0   1   0   2   0   0   0   0   0
#> 12 341 331 322  0  0  0  0  0  0   0   0   1   1   0   1   0   0   2   0   0
#> 13 357 315 321  0  1  0  0  1  0   0   0   1   0   1   0   0   0   0   1   0
#> 14 283 369 319  2  4  0  2  1  0   3   2   2   1   0   1   0   1   1   1   0
#> 15 267 289 282  4  6  6  6  4  8   8   5   7   3  10  10   5   5   4   2   5
#> 16 134 143 153 25 19 15 26 19 28  21  22  30  13  16  22  18  17  16  23  21
#> 17  25  26  36 31 30 29 33 41 23  41  30  35  30  33  34  26  35  28  40  53
#> 18   6   2   5 36 31 32 38 37 38  24  41  50  46  39  32  30  35  27  40  29
#> 19   0   1   0 41 35 34 39 22 44  44  31  44  41  39  36  51  26  25  34  30
#> 20   0   0   0 30 29 36 41 27 29  44  32  33  31  48  30  31  51  37  37  35
#> 21   3   0   0 30 42 32 28 44 41  39  43  36  42  42  31  25  36  42  42  38
#> 22   7  12   9 39 36 41 48 40 34  42  31  40  30  36  29  45  38  29  42  35
#> 23  15  23  16 35 30 39 40 35 35  30  34  34  41  32  21  38  34  29  35  38
#> 24  15  29  21 24 48 45 36 29 35  33  19  39  32  36  44  51  35  41  35  31
#> 25  25  32  27 24 38 30 24 41 28  32  34  35  37  31  31  41  30  31  34  33
#> 26  29  30  37 35 28 30 32 23 34  31  27  39  27  36  26  36  44  29  40  41
#> 27  44  40  31 32 25 46 37 35 24  35  24  30  45  28  34  30  34  29  21  29
#> 28  31  23  32 31 36 39 41 40 27  36  26  26  38  30  29  33  33  38  31  41
#> 29  41  30  33 36 38 39 38 33 33  28  42  27  26  35  22  38  22  29  35  36
#> 30  32  31  39 31 43 35 36 29 44  31  26  32  29  37  27  26  28  36  36  30
#>    V21 V22 V23 V24 V25 V26 V27 V28 V29 V30
#> 1   72  16  28   4  30  27  44  59   6  21
#> 2   51  21  30  11  33  26  44  60  15  28
#> 3   44  24  31  19  29  32  36  29  19  35
#> 4   19  24  34  29  26  30  23  31  23  29
#> 5   28  24  25  23  25  27  27  27  33  35
#> 6   14  22  17  14  11  13  12  23  14  23
#> 7    9   1   1   9   4   7   4   3   4   3
#> 8    0   2   1   3   2   0   2   0   0   0
#> 9    0   0   0   0   1   1   0   0   1   0
#> 10   0   0   0   1   0   0   1   0   0   0
#> 11   0   0   0   0   0   0   0   0   0   0
#> 12   0   0   0   0   1   0   0   0   0   0
#> 13   0   0   0   0   0   1   0   0   1   0
#> 14   2   1   3   1   0   0   1   0   0   0
#> 15   4   3  13   6   9   6   9   3   8   3
#> 16  16  18  15  28  18  20  27  25  25  27
#> 17  28  36  33  37  47  30  31  31  34  34
#> 18  39  52  41  36  35  31  31  37  44  36
#> 19  37  36  40  43  43  34  43  46  25  36
#> 20  34  38  43  37  35  32  49  54  37  40
#> 21  38  46  30  38  36  36  35  39  35  31
#> 22  40  35  26  31  38  27  33  35  34  38
#> 23  42  36  32  43  34  35  42  39  35  28
#> 24  29  34  37  26  36  32  39  31  28  30
#> 25  39  39  40  36  32  26  37  42  39  32
#> 26  31  32  41  32  34  25  38  27  41  45
#> 27  29  28  30  37  35  32  42  41  34  39
#> 28  33  38  37  31  38  30  33  34  30  35
#> 29  28  40  28  30  40  34  31  35  33  40
#> 30  38  23  26  27  30  29  49  38  39  43