Simulate microbiome perturbation data

To simulate microbiome longitudinal count data for perturbation experiments with user defined starting and ending time point, effect size and duration of perturbation

sim_ts_perturb(
  endtime,
  n_species,
  n_depth,
  Sigma = NULL,
  beta,
  t1,
  t2,
  time,
  species,
  duration
)

Arguments

endtime

the number of time points of the sample

n_species

number of species

n_depth

sequence depth

Sigma

correlation matrix for muti-nomal distribution

beta

effect of perturbation

t1

time of reaching the maximum of perturbation

t2

time of ending the maximum of perturbation

time

the start time for perturbation

species

perturbed species

duration

the duration of perturbation

Value

the simulated data

Examples

sim_ts_perturb(endtime=30, n_species=30, 1000, Sigma=NULL, beta=1,species=c(1,2,3),time=4, t1=8,t2=12,duration = 16)
#>     V1  V2  V3 V4 V5 V6 V7 V8 V9 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V20
#> 1   37  23  40 27 46 23  6 22 26  51  23  50  38  55  52  25  65  44  13  37
#> 2   35  27  28 20 58 28 18 21 25  35  43  50  26  49  36  27  35  26  34  42
#> 3   42  26  30 26 49 37 12 38 25  36  26  43  31  37  37  31  53  28  26  36
#> 4   31  31  33 45 30 38 29 35 25  29  22  39  30  35  53  25  40  28  45  38
#> 5   75  76  74 30 25 29 29 31 28  29  22  29  26  36  31  31  28  31  29  28
#> 6  198 199 186 13 21 16 14 13 16  21  20  13  11  15  15  22  13  13  17  16
#> 7  293 283 300  5  4  2  5  5  4   8   3   4   5   2   4   5   6   4   4   8
#> 8  337 347 306  0  0  0  0  1  0   1   0   0   1   0   0   0   0   1   0   0
#> 9  341 347 303  0  0  0  1  0  0   2   1   1   0   1   0   0   0   0   0   0
#> 10 342 346 307  0  0  0  0  1  0   0   0   0   1   0   0   0   0   0   0   0
#> 11 319 339 339  0  0  0  0  0  0   0   0   0   1   0   0   0   0   1   0   0
#> 12 311 341 348  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0
#> 13 329 325 340  0  0  0  0  0  0   1   0   0   1   0   1   0   1   0   0   0
#> 14 300 315 346  4  2  1  0  1  0   0   0   1   0   2   1   4   1   1   0   4
#> 15 254 281 264  3 11  3  7 12 10   6   8   8   2   8   9   5   5   8   5   6
#> 16 145 126 132 31 23 24 31 29 25  18  18  21  16  23  21  24  26  24  17  16
#> 17  35  24  32 33 38 36 39 18 32  34  28  29  35  45  30  33  39  25  42  36
#> 18   6   6   3 49 36 36 45 47 20  36  29  37  38  30  41  42  31  45  24  43
#> 19   0   1   0 39 34 40 35 35 41  41  25  38  34  35  43  42  36  42  45  44
#> 20   0   0   0 32 50 33 45 33 41  44  36  37  35  41  40  33  35  40  35  28
#> 21   2   2   0 32 28 42 42 30 30  39  34  38  45  39  39  42  29  38  46  34
#> 22   7   9  11 34 41 34 24 37 37  45  37  33  41  45  39  34  38  22  32  33
#> 23  19  21  21 31 44 30 33 31 38  37  37  34  27  35  35  35  35  46  31  32
#> 24  23  26  27 25 42 38 38 26 28  35  41  43  33  42  30  31  32  39  28  32
#> 25  26  28  20 30 35 32 37 23 38  29  33  32  35  34  42  38  36  36  31  35
#> 26  18  27  41 32 29 36 56 36 37  28  33  32  32  27  35  34  36  40  31  29
#> 27  36  41  38 24 32 30 35 34 42  37  32  37  30  37  39  33  33  39  31  26
#> 28  33  24  32 44 26 28 33 30 28  36  37  28  26  34  37  45  37  41  27  34
#> 29  29  28  32 31 28 32 38 35 40  27  31  42  32  28  48  38  39  23  32  26
#> 30  32  32  30 41 26 41 22 28 31  35  31  40  38  39  43  23  30  34  33  28
#>    V21 V22 V23 V24 V25 V26 V27 V28 V29 V30
#> 1   77  23  29  23  18  40  23  31  10  23
#> 2   37  36  37  29  28  35  27  40  30  38
#> 3   44  25  31  27  28  43  38  33  31  31
#> 4   38  26  25  30  33  33  36  41  30  27
#> 5   30  33  31  34  21  31  27  32  19  25
#> 6   13  15  11  15  14  14  22  12  14  18
#> 7    4   5   4   1   4   7   5   8   3   5
#> 8    0   0   1   1   0   3   0   1   0   0
#> 9    0   0   0   0   0   0   0   1   0   2
#> 10   0   1   0   0   1   0   0   0   1   0
#> 11   0   0   0   0   0   0   0   1   0   0
#> 12   0   0   0   0   0   0   0   0   0   0
#> 13   1   0   0   0   0   0   0   0   0   1
#> 14   1   2   2   3   1   2   3   2   0   1
#> 15   5  10   5  11  13  15   6   6   7   7
#> 16  26  21  21  21  19  13  20  23  25  21
#> 17  46  29  32  36  29  36  36  31  32  30
#> 18  30  43  35  48  36  30  35  44  27  28
#> 19  33  33  36  38  31  40  33  34  37  35
#> 20  25  52  35  34  43  39  29  40  29  36
#> 21  33  45  40  34  44  29  31  37  41  35
#> 22  36  32  42  32  25  39  43  33  45  40
#> 23  24  45  40  41  31  40  24  45  29  29
#> 24  39  31  34  29  34  33  39  37  26  39
#> 25  32  20  46  51  36  21  37  41  28  38
#> 26  36  28  28  38  37  30  36  24  43  31
#> 27  22  32  34  24  37  37  34  29  33  32
#> 28  39  31  28  28  37  34  23  41  39  40
#> 29  31  30  36  33  40  32  34  43  32  30
#> 30  31  34  39  39  26  32  33  40  41  28