Simulate microbiome perturbation data

To simulate microbiome longitudinal count data for perturbation experiments with user defined starting and ending time point, effect size and duration of perturbation

sim_ts_perturb(
  endtime,
  n_species,
  n_depth,
  Sigma = NULL,
  beta,
  t1,
  t2,
  time,
  species,
  duration
)

Arguments

endtime

the number of time points of the sample

n_species

number of species

n_depth

sequence depth

Sigma

correlation matrix for muti-nomal distribution

beta

effect of perturbation

t1

time of reaching the maximum of perturbation

t2

time of ending the maximum of perturbation

time

the start time for perturbation

species

perturbed species

duration

the duration of perturbation

Value

the simulated data

Examples

sim_ts_perturb(endtime=30, n_species=30, 1000, Sigma=NULL, beta=1,species=c(1,2,3),time=4, t1=8,t2=12,duration = 16)
#>     V1  V2  V3 V4 V5 V6 V7 V8 V9 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V20
#> 1   34  55  16 28  8 20 83 12 40  23  46  36  36  65  10  40  29  41  22  29
#> 2   32  37  25 35 15 23 58 21 49  29  36  33  39  48  24  38  34  37  16  39
#> 3   35  48  21 37 18 26 46 18 41  33  40  33  32  39  30  38  37  27  26  42
#> 4   38  30  29 33 18 36 36 31 28  41  30  34  35  37  28  34  32  38  36  42
#> 5   75  77  66 28 24 24 26 24 28  28  31  32  29  32  32  43  20  24  32  37
#> 6  196 188 170 16 17 15 19  6 16  15  14  19  20  16  22  13  13  17  14  19
#> 7  316 313 267  7  4  4  5  3  4   3   2   3   3   3   5   3   5   6   5   6
#> 8  317 339 324  2  1  1  1  1  0   2   1   1   0   0   0   2   2   0   0   2
#> 9  323 352 319  0  0  0  1  0  0   1   1   0   0   0   0   0   0   0   0   1
#> 10 336 323 337  0  0  0  0  0  0   0   0   0   1   0   0   0   0   0   0   1
#> 11 330 335 332  0  0  0  0  0  1   0   1   0   0   0   0   0   0   0   0   1
#> 12 324 341 335  0  0  0  0  0  0   0   0   0   0   0   0   0   0   0   0   0
#> 13 292 361 343  0  0  0  0  0  0   1   0   0   0   0   0   0   0   0   2   0
#> 14 349 303 305  3  1  4  3  1  0   1   0   0   0   1   2   1   3   5   1   2
#> 15 275 304 275  6  4  6  4  4  9   9   5   6   1   7   4   4   7   5   5   9
#> 16 136 142 172 18 18 17 19 21 31  22  18  20  24  27  18  17  13  25  17   9
#> 17  30  35  24 28 36 40 35 32 31  36  23  43  31  35  37  32  30  25  36  39
#> 18   3   4   8 38 40 41 30 40 35  30  47  37  37  37  33  25  29  44  39  42
#> 19   2   2   0 43 39 27 42 37 31  37  49  29  30  33  37  32  47  37  38  43
#> 20   0   0   0 35 33 30 25 44 37  32  38  35  46  34  40  43  46  32  42  37
#> 21   3   2   1 36 53 24 45 40 42  39  27  34  45  42  40  36  37  39  42  29
#> 22   6   4   8 31 48 42 29 38 36  39  36  34  53  26  33  51  41  41  28  34
#> 23  24  15  15 28 34 43 45 28 32  38  34  29  40  40  32  42  33  25  29  48
#> 24  25  34  23 28 28 35 33 35 37  27  40  38  26  30  37  31  28  34  36  30
#> 25  21  24  40 32 30 36 43 37 32  34  41  34  28  32  36  44  29  31  30  23
#> 26  33  34  32 38 39 25 38 31 27  34  30  33  37  45  29  30  31  37  35  30
#> 27  32  38  31 41 32 43 38 31 40  44  24  28  36  39  32  27  33  31  33  30
#> 28  35  31  33 36 32 29 44 45 29  45  32  28  40  32  34  41  32  28  30  24
#> 29  52  28  35 25 38 29 34 31 32  24  27  34  32  18  48  25  33  44  22  39
#> 30  41  29  37 45 32 23 39 25 33  33  40  29  27  39  34  38  20  34  34  32
#>    V21 V22 V23 V24 V25 V26 V27 V28 V29 V30
#> 1   33  27  49  21  27  62  35  17  28  28
#> 2   30  28  39  35  29  47  30  26  31  37
#> 3   25  38  38  20  22  50  38  25  37  40
#> 4   27  33  50  36  32  27  26  30  39  34
#> 5   31  27  30  34  27  27  34  29  25  24
#> 6   23  18  13  17  17  21  15  18  19  14
#> 7    0   2   5   1   4   4   1   6   7   3
#> 8    2   0   0   0   1   0   0   1   0   0
#> 9    0   0   1   0   1   0   0   0   0   0
#> 10   0   0   0   0   1   0   0   0   1   0
#> 11   0   0   0   0   0   0   0   0   0   0
#> 12   0   0   0   0   0   0   0   0   0   0
#> 13   0   1   0   0   0   0   0   0   0   0
#> 14   0   2   3   1   3   2   0   1   0   3
#> 15   7   6   8   3   5   4   6   3   5   4
#> 16  24  30  26  25  18  19  17  17  21  19
#> 17  32  41  34  41  25  32  32  27  48  30
#> 18  50  34  38  36  30  26  36  35  35  41
#> 19  42  31  41  31  44  28  37  41  40  30
#> 20  42  36  32  38  42  43  29  32  42  35
#> 21  40  38  29  43  28  31  40  27  30  38
#> 22  30  34  41  25  33  42  38  34  35  30
#> 23  36  37  40  34  30  28  29  32  49  31
#> 24  43  34  37  38  43  40  39  28  35  28
#> 25  35  34  27  32  28  41  43  32  38  33
#> 26  34  25  35  38  29  46  38  18  31  38
#> 27  28  42  24  34  35  33  28  25  42  26
#> 28  34  29  33  34  38  38  31  34  27  22
#> 29  34  32  32  30  39  40  41  35  29  38
#> 30  35  31  37  30  32  30  38  28  36  39