Simulate microbiome perturbation data

To simulate microbiome longitudinal count data for perturbation experiments with user defined starting and ending time point, effect size and duration of perturbation

sim_ts_perturb(
  endtime,
  n_species,
  n_depth,
  Sigma = NULL,
  beta,
  t1,
  t2,
  time,
  species,
  duration
)

Arguments

endtime

the number of time points of the sample

n_species

number of species

n_depth

sequence depth

Sigma

correlation matrix for muti-nomal distribution

beta

effect of perturbation

t1

time of reaching the maximum of perturbation

t2

time of ending the maximum of perturbation

time

the start time for perturbation

species

perturbed species

duration

the duration of perturbation

Value

the simulated data

Examples

sim_ts_perturb(endtime=30, n_species=30, 1000, Sigma=NULL, beta=1,species=c(1,2,3),time=4, t1=8,t2=12,duration = 16)
#>     V1  V2  V3 V4 V5  V6 V7 V8 V9 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V20
#> 1    6  14  13 78 33 127 31 16 11  25   9  34  25  11  58 128  59  19  15  44
#> 2   26  23  23 65 45  58 36 15 35  29  14  36  27  23  41  72  47  26  34  44
#> 3   25  34  49 49 39  48 32 30 31  21  29  31  37  19  51  40  26  30  31  33
#> 4   31  21  29 46 33  41 31 34 27  23  41  28  31  32  48  39  43  25  22  40
#> 5   71  72  81 39 30  32 32 28 29  31  21  22  17  19  40  30  38  30  29  27
#> 6  184 181 185 13 28  17 19 22 13  11  15   9  17  16  14  24  20  14   8  18
#> 7  286 290 313  5  5   1  5  5  4   5   5   3   4   4   8   5   2   3   3   2
#> 8  325 335 321  1  2   0  0  1  1   0   2   0   2   0   0   3   0   0   2   0
#> 9  318 326 351  0  0   1  1  0  0   0   0   0   0   0   0   0   1   0   0   0
#> 10 332 330 328  0  0   2  1  1  0   1   0   0   0   1   1   1   0   0   0   0
#> 11 314 356 325  1  0   0  1  0  0   0   0   0   0   1   0   0   0   1   0   0
#> 12 305 332 359  0  0   0  0  0  0   1   0   0   0   0   0   1   0   1   0   0
#> 13 322 348 320  0  0   1  1  0  0   0   0   0   0   0   1   1   1   1   1   0
#> 14 329 301 324  0  2   2  1  5  2   0   1   0   1   2   3   4   4   1   1   2
#> 15 284 249 285  9  7   6  6  6  7   4   5   3  10  10   9   4   6   8   6   4
#> 16 132 128 121 29 23  27 23 21 22  20  18  23  23  23  18  25  25  18  18  23
#> 17  34  32  35 28 26  26 28 37 27  42  31  42  33  44  27  31  27  43  28  20
#> 18   3   3   9 43 24  30 30 35 42  39  43  37  34  38  33  38  32  39  40  31
#> 19   0   2   0 37 34  42 43 44 44  38  30  44  40  36  26  32  32  39  41  38
#> 20   0   0   0 41 52  48 34 30 37  34  43  46  33  31  36  28  46  39  31  29
#> 21   1   2   3 34 29  37 41 37 30  32  38  38  46  30  30  46  56  36  39  36
#> 22   8   6  10 34 36  32 37 40 38  38  32  32  33  35  28  32  25  33  38  43
#> 23  10  19  15 35 31  35 40 35 36  33  27  39  38  31  39  41  36  31  31  27
#> 24  24  22  23 40 28  40 32 40 37  28  41  39  35  34  28  39  31  44  28  40
#> 25  29  25  19 39 32  49 39 34 33  31  44  35  35  34  25  28  34  36  32  38
#> 26  36  34  21 26 36  36 39 36 34  38  44  42  42  35  25  27  29  28  32  37
#> 27  35  26  25 24 31  25 28 38 44  31  32  33  36  37  25  22  35  41  48  32
#> 28  29  27  30 35 34  29 40 23 39  25  27  29  40  37  39  42  43  35  39  32
#> 29  33  28  32 37 16  33 36 36 54  41  36  21  34  29  31  35  24  30  28  44
#> 30  31  19  29 38 38  38 32 23 42  34  37  29  35  25  45  31  27  39  27  41
#>    V21 V22 V23 V24 V25 V26 V27 V28 V29 V30
#> 1   24   6  29  35  20  30  37  10  17  36
#> 2   22  11  30  32  31  30  35  27  26  37
#> 3   21  33  28  38  33  36  43  24  30  29
#> 4   29  36  39  36  27  30  39  31  31  37
#> 5   35  31  21  35  25  21  29  34  20  31
#> 6   15  14  18  18  19  17  18  14  19  20
#> 7    3   6   4   8   5   3   3   5   3   2
#> 8    0   0   0   1   0   1   1   1   0   1
#> 9    0   0   0   0   0   1   1   0   0   0
#> 10   0   0   1   0   0   1   0   0   0   0
#> 11   0   0   0   0   0   0   0   0   1   0
#> 12   1   0   0   0   0   0   0   0   0   0
#> 13   0   0   1   0   0   1   0   0   1   0
#> 14   1   1   1   3   1   0   3   3   2   0
#> 15  10   5   4  10  11   7   6  11   3   5
#> 16  25  17  25  28  31  22  22  21  30  19
#> 17  40  39  29  41  36  37  33  33  36  35
#> 18  37  36  36  36  39  57  30  36  42  28
#> 19  28  44  37  39  43  29  34  35  32  37
#> 20  34  32  39  36  24  36  36  40  55  30
#> 21  47  36  30  28  30  38  37  40  37  36
#> 22  37  43  34  43  37  40  37  40  38  41
#> 23  39  26  41  34  40  41  45  35  33  37
#> 24  25  30  27  38  38  42  29  37  36  25
#> 25  23  29  26  29  39  37  32  39  42  33
#> 26  33  28  43  33  24  36  28  32  36  30
#> 27  36  29  41  47  24  35  38  29  40  33
#> 28  30  37  35  38  35  33  29  38  30  21
#> 29  44  33  36  25  23  29  43  33  38  38
#> 30  35  46  34  24  34  43  44  26  27  27