Simulate microbiome perturbation data

To simulate microbiome longitudinal count data for perturbation experiments with user defined starting and ending time point, effect size and duration of perturbation

sim_ts_perturb(
  endtime,
  n_species,
  n_depth,
  Sigma = NULL,
  beta,
  t1,
  t2,
  time,
  species,
  duration
)

Arguments

endtime

the number of time points of the sample

n_species

number of species

n_depth

sequence depth

Sigma

correlation matrix for muti-nomal distribution

beta

effect of perturbation

t1

time of reaching the maximum of perturbation

t2

time of ending the maximum of perturbation

time

the start time for perturbation

species

perturbed species

duration

the duration of perturbation

Value

the simulated data

Examples

sim_ts_perturb(endtime=30, n_species=30, 1000, Sigma=NULL, beta=1,species=c(1,2,3),time=4, t1=8,t2=12,duration = 16)
#>     V1  V2  V3 V4 V5 V6 V7 V8 V9 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V20
#> 1   21  20  22 22 29 11 25 17 19  25  36  21  57  50  13  23  32  16  51  41
#> 2   15  29  36 22 22 14 27 15 40  30  52  30  50  50  31  34  35  22  40  36
#> 3   38  32  20 41 30 22 28 31 31  31  34  31  35  32  34  25  32  31  36  38
#> 4   20  34  28 33 26 29 33 25 28  30  35  24  34  29  30  47  34  39  34  35
#> 5   81  56  82 23 29 27 32 26 25  30  25  27  32  34  31  29  31  19  39  35
#> 6  205 170 198  9 14 13 11 14 18  19  16  23  21  19  17  20  11  11  24  12
#> 7  305 273 292  6  4  6  3  2  5   3   6   7   4   5   4   4   3   5   5   4
#> 8  362 296 324  1  0  1  3  2  0   1   0   0   1   1   1   0   2   1   1   0
#> 9  341 322 330  0  0  0  0  0  0   0   2   0   0   0   0   0   1   0   0   1
#> 10 322 331 339  0  0  1  0  0  0   0   0   0   2   1   0   1   0   1   0   0
#> 11 298 362 335  0  0  0  1  0  0   0   0   0   1   0   0   0   0   1   0   1
#> 12 326 337 335  0  0  0  0  0  0   0   0   0   0   0   0   0   1   0   0   0
#> 13 335 300 355  0  0  0  2  1  0   0   0   0   1   0   0   0   0   1   0   1
#> 14 315 304 343  2  1  1  2  2  0   3   2   2   4   1   2   1   0   1   0   3
#> 15 262 235 312  6  6  7  6  6  6   9   6   6  14   8   7   3   7  10   3  13
#> 16 139 138 143 25 17 15 25 27 31  24  21  21  31  17  16  14  22  24  26  20
#> 17  37  31  27 37 28 33 32 24 32  29  27  36  50  29  30  32  40  38  34  33
#> 18   5   3   4 37 37 31 36 33 32  28  36  37  32  43  40  35  49  33  39  43
#> 19   2   0   0 42 43 37 41 37 43  32  52  31  32  35  30  35  37  32  32  37
#> 20   0   0   0 23 33 46 36 43 40  39  41  32  32  35  39  43  35  28  34  36
#> 21   1   1   2 28 42 33 31 50 38  42  36  36  33  24  40  25  39  30  27  36
#> 22   9   8   7 40 42 39 36 31 33  27  26  39  49  38  24  43  41  40  40  34
#> 23  15  13  27 30 29 57 40 37 47  32  30  28  33  34  38  29  35  41  45  34
#> 24  24  30  34 37 33 18 27 35 35  22  39  48  36  29  36  38  34  33  32  40
#> 25  26  30  26 32 37 40 41 38 34  25  37  36  37  32  29  38  43  28  31  27
#> 26  36  36  21 39 38 33 35 41 41  31  34  21  35  33  38  32  28  36  31  29
#> 27  25  27  35 23 47 33 28 32 36  34  27  44  29  38  41  25  38  33  31  38
#> 28  28  30  36 29 35 39 29 33 27  29  35  37  36  43  31  36  30  41  35  34
#> 29  38  43  24 35 32 42 29 38 39  34  31  37  28  32  34  26  29  33  31  35
#> 30  23  26  28 37 21 30 36 33 37  46  35  32  37  34  33  36  28  33  35  36
#>    V21 V22 V23 V24 V25 V26 V27 V28 V29 V30
#> 1   46  36  25  46  44  64  32  32  82  42
#> 2   39  43  31  32  32  47  35  30  51  30
#> 3   31  38  30  42  36  45  28  37  39  42
#> 4   39  40  41  36  46  28  36  41  42  24
#> 5   28  24  25  24  30  41  24  34  31  26
#> 6   17  16  14  16  29  13  14  15   9  12
#> 7    5   6  12   6   2   6   2   6   4   5
#> 8    2   0   0   0   0   0   0   0   1   0
#> 9    0   0   1   0   0   1   0   0   1   0
#> 10   1   0   0   0   0   0   0   0   0   1
#> 11   0   1   0   0   0   0   0   0   0   0
#> 12   0   0   0   0   0   1   0   0   0   0
#> 13   0   0   1   0   1   2   0   0   0   0
#> 14   2   2   2   0   2   0   0   1   0   2
#> 15   4   7   1   4   5  15   8   4  10  10
#> 16  15  19  19  21  18  15  23  18  33  23
#> 17  30  39  36  42  33  31  35  29  35  31
#> 18  42  32  44  43  30  35  35  38  29  39
#> 19  34  44  39  28  38  53  31  38  28  37
#> 20  44  40  41  35  35  38  36  43  42  31
#> 21  40  39  34  47  38  53  41  36  37  41
#> 22  40  35  38  34  30  41  41  35  30  30
#> 23  34  34  48  24  26  36  31  30  28  35
#> 24  16  39  36  37  24  42  33  43  29  41
#> 25  34  40  27  25  36  31  28  34  38  40
#> 26  35  41  32  43  27  30  29  25  32  38
#> 27  28  33  36  35  36  35  33  43  35  22
#> 28  38  34  34  34  39  26  33  35  23  31
#> 29  25  37  31  20  31  38  39  26  39  44
#> 30  31  39  28  34  38  39  33  43  24  35