Simulate microbiome perturbation data

To simulate microbiome longitudinal count data for perturbation experiments with user defined starting and ending time point, effect size and duration of perturbation

sim_ts_perturb(
  endtime,
  n_species,
  n_depth,
  Sigma = NULL,
  beta,
  t1,
  t2,
  time,
  species,
  duration
)

Arguments

endtime

the number of time points of the sample

n_species

number of species

n_depth

sequence depth

Sigma

correlation matrix for muti-nomal distribution

beta

effect of perturbation

t1

time of reaching the maximum of perturbation

t2

time of ending the maximum of perturbation

time

the start time for perturbation

species

perturbed species

duration

the duration of perturbation

Value

the simulated data

Examples

sim_ts_perturb(endtime=30, n_species=30, 1000, Sigma=NULL, beta=1,species=c(1,2,3),time=4, t1=8,t2=12,duration = 16)
#>     V1  V2  V3 V4 V5 V6 V7 V8 V9 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V20
#> 1   35  27  42 51 34 19 24 37 32  12  36  13  44  62  22  41  67  26  31  35
#> 2   39  30  40 36 25 18 28 36 36  21  25  27  41  50  31  32  42  31  48  30
#> 3   38  27  34 49 35 30 36 41 28  28  25  35  37  37  27  40  32  41  28  31
#> 4   34  44  39 25 38 23 28 35 41  21  34  31  43  37  32  30  43  37  37  33
#> 5   92  68  77 29 39 29 18 29 23  31  32  20  28  23  35  25  39  23  31  27
#> 6  217 194 190 15 17 15 13 16 21  12  12  11  15  12  18  17  12  14  17  21
#> 7  296 307 297  1  4  4  1  8  7   6   4   2   6   2   3   3   4   3   4   2
#> 8  322 340 319  2  1  1  2  0  0   0   1   1   0   2   1   1   0   3   0   0
#> 9  317 344 330  1  1  1  0  0  0   0   1   2   0   0   0   0   0   0   0   0
#> 10 305 372 319  0  0  0  1  0  1   0   0   0   0   0   0   0   0   1   0   0
#> 11 306 346 343  0  0  0  0  0  0   0   0   0   0   0   1   0   1   0   1   0
#> 12 323 331 343  0  0  0  0  0  0   0   0   0   0   0   1   0   0   0   1   0
#> 13 315 334 343  0  0  1  1  0  0   0   1   0   0   0   1   0   1   0   0   0
#> 14 304 319 329  2  2  3  1  1  1   2   1   1   1   3   1   3   2   1   2   3
#> 15 277 288 284  2  5 13  2  0  5   6   8   7   7  12  11   4   5   6   4   5
#> 16 154 151 128 22 18 18 22 24 16  23  26  28  18  21  25  22  19  13  17  21
#> 17  25  26  28 30 26 34 25 35 32  28  48  41  25  50  36  32  31  33  28  27
#> 18   6   4   3 35 39 32 34 28 42  34  36  31  37  28  36  35  35  41  32  50
#> 19   1   0   1 33 39 40 47 42 34  30  29  26  35  40  42  39  47  41  32  41
#> 20   0   0   0 40 40 35 39 37 33  39  38  32  47  35  34  43  54  42  27  38
#> 21   0   2   1 35 44 41 36 44 40  40  27  38  41  32  47  36  46  36  39  35
#> 22   2   9   9 34 38 35 28 30 33  39  28  41  38  50  28  41  42  38  38  41
#> 23  16  28  16 35 34 42 39 31 31  42  32  30  35  32  33  35  34  28  43  41
#> 24  23  28  13 40 30 35 26 34 44  38  31  26  31  36  31  33  26  41  34  32
#> 25  32  33  22 27 38 28 29 46 36  37  33  29  25  33  36  41  33  44  31  34
#> 26  34  34  38 39 22 36 32 40 35  23  33  23  30  35  36  19  39  38  28  40
#> 27  47  28  32 35 33 36 28 24 33  46  29  31  35  30  37  27  36  33  38  37
#> 28  36  27  18 46 40 28 40 43 26  30  32  29  28  30  41  28  45  40  26  26
#> 29  24  27  27 28 30 38 37 38 34  35  43  27  35  31  37  34  38  39  22  30
#> 30  28  33  40 34 34 30 33 35 38  25  30  32  31  34  29  40  30  31  34  34
#>    V21 V22 V23 V24 V25 V26 V27 V28 V29 V30
#> 1   24  22  12  47  14  39  75  23  28  26
#> 2   30  32  18  50  30  35  61  27  19  32
#> 3   25  28  24  35  29  36  34  36  33  41
#> 4   31  24  24  36  39  31  33  27  33  37
#> 5   26  23  28  21  32  35  29  35  27  26
#> 6    6  16  20  16  11  14  16  12  15  15
#> 7    5   1   5   1   1   3   5   6   4   5
#> 8    0   0   1   0   0   0   0   2   1   0
#> 9    0   1   0   1   0   1   0   0   0   0
#> 10   0   0   1   0   0   0   0   0   0   0
#> 11   0   0   0   0   0   1   0   0   1   0
#> 12   0   0   0   0   0   0   0   0   1   0
#> 13   0   0   0   1   0   1   1   0   0   0
#> 14   2   2   2   2   2   1   0   3   2   2
#> 15   4   7   4   5   2   8   5   5   6   3
#> 16  24  17  20  26  14  20  24  23  26  20
#> 17  43  27  41  31  30  33  44  38  37  36
#> 18  33  35  49  37  35  33  37  41  45  37
#> 19  41  31  37  33  35  42  24  46  35  37
#> 20  27  29  36  34  39  44  39  40  31  28
#> 21  33  34  34  31  31  48  41  27  28  33
#> 22  37  43  32  36  36  43  41  29  34  27
#> 23  41  37  39  31  30  35  30  29  36  35
#> 24  39  42  46  30  29  40  35  32  40  35
#> 25  41  42  41  22  39  35  31  29  28  25
#> 26  32  46  41  25  26  35  33  39  28  41
#> 27  31  39  29  33  25  31  30  34  37  36
#> 28  32  39  23  35  30  37  35  47  33  30
#> 29  34  51  31  34  25  46  28  31  30  36
#> 30  30  36  37  37  39  37  28  30  32  39