Simulate microbiome perturbation data

To simulate microbiome longitudinal count data for perturbation experiments with user defined starting and ending time point, effect size and duration of perturbation

sim_ts_perturb(
  endtime,
  n_species,
  n_depth,
  Sigma = NULL,
  beta,
  t1,
  t2,
  time,
  species,
  duration
)

Arguments

endtime

the number of time points of the sample

n_species

number of species

n_depth

sequence depth

Sigma

correlation matrix for muti-nomal distribution

beta

effect of perturbation

t1

time of reaching the maximum of perturbation

t2

time of ending the maximum of perturbation

time

the start time for perturbation

species

perturbed species

duration

the duration of perturbation

Value

the simulated data

Examples

sim_ts_perturb(endtime=30, n_species=30, 1000, Sigma=NULL, beta=1,species=c(1,2,3),time=4, t1=8,t2=12,duration = 16)
#>     V1  V2  V3 V4 V5 V6 V7 V8 V9 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V20
#> 1   19   5  40 12 59 37 46 50 26  12  23  57  17  17 100  69  39  24  42  16
#> 2   30  24  38 22 40 44 40 39 29  17  31  44  27  23  51  52  43  33  42  20
#> 3   34  29  26 23 36 35 38 31 39  16  35  44  28  39  45  32  34  26  34  37
#> 4   36  24  42 26 37 38 24 39 31  39  34  37  37  26  35  34  32  26  33  31
#> 5   87  66  90 35 30 29 25 32 23  25  25  28  20  28  30  23  36  29  33  27
#> 6  174 203 198 13 18 16 19 16 13  15  17  14  11  13  16  19  17  17  13  22
#> 7  309 287 292  4  5  4  5  3  1   4   4   5   6   4   6   3   5   4   2   3
#> 8  334 330 321  0  1  0  0  0  1   0   1   1   2   0   1   0   0   2   0   0
#> 9  326 335 331  1  0  0  1  1  0   0   0   0   0   0   0   1   0   0   0   0
#> 10 320 334 338  0  0  1  1  0  1   0   0   0   0   0   1   1   0   0   0   1
#> 11 342 312 343  0  0  1  0  0  0   0   0   0   0   0   0   0   0   0   0   0
#> 12 298 366 332  0  0  0  0  1  0   0   0   1   0   0   0   0   0   0   0   0
#> 13 307 346 338  0  0  1  0  0  0   0   0   1   0   0   0   0   0   1   0   1
#> 14 319 306 334  0  1  0  1  3  5   2   1   3   2   0   1   0   1   0   2   0
#> 15 272 261 290  8  8  7  3  8  3   6   9   3   6   9   4   7  10  11   2   4
#> 16 139 149 153 13 30 30 29 21 24  23  23  22  19  22  20  19  14  30  21  12
#> 17  40  28  29 28 33 30 28 31 28  31  25  41  39  30  34  31  28  37  31  33
#> 18   5   5   8 41 38 34 37 35 32  38  41  39  34  37  31  29  45  24  37  43
#> 19   1   1   3 34 41 38 37 49 28  39  39  37  33  30  37  37  32  31  31  43
#> 20   0   0   0 44 30 32 38 27 40  30  51  41  40  42  25  45  43  45  27  31
#> 21   1   0   2 35 37 42 29 35 34  29  38  38  39  42  39  32  37  31  44  36
#> 22   9  14   5 30 29 40 29 39 42  33  39  24  28  39  36  43  38  36  37  39
#> 23  16  15  19 42 32 28 37 39 37  33  38  36  34  30  48  34  33  36  42  39
#> 24  22  35  13 29 30 28 34 31 39  34  39  36  32  40  42  28  29  44  28  38
#> 25  25  32  23 40 36 25 40 32 30  28  35  27  29  38  36  30  37  30  41  37
#> 26  25  24  33 33 43 37 36 23 38  27  37  21  38  31  39  28  24  45  29  29
#> 27  26  37  29 42 37 29 29 40 29  33  35  29  31  35  29  28  37  35  38  27
#> 28  40  27  30 34 31 35 33 26 27  36  44  32  31  35  34  36  38  38  28  34
#> 29  33  35  38 34 43 32 31 36 32  31  27  32  25  36  39  43  29  36  38  26
#> 30  24  23  29 32 35 41 36 31 39  33  38  35  45  35  32  35  31  36  35  19
#>    V21 V22 V23 V24 V25 V26 V27 V28 V29 V30
#> 1   49  13  34  20  19  17  34  34  43  27
#> 2   43  22  24  29  27  20  45  30  40  31
#> 3   27  33  23  43  23  39  33  45  36  37
#> 4   37  29  33  32  36  34  31  32  36  39
#> 5   26  32  17  32  28  27  28  26  33  30
#> 6   28  15  15  12  15  15  12  16  13  15
#> 7    1   6   7   5   3   2   8   2   3   7
#> 8    0   0   1   0   1   0   1   2   1   0
#> 9    0   0   0   1   1   0   1   0   0   1
#> 10   0   0   1   0   0   0   0   1   0   0
#> 11   0   1   0   1   0   0   0   0   0   0
#> 12   1   0   0   0   0   0   0   0   0   1
#> 13   1   1   0   0   0   0   0   2   1   0
#> 14   3   1   3   4   0   2   1   3   1   1
#> 15   7   4   7   6   4   5   9  14   7   6
#> 16  20  19  25  20  18  14  15  19  20  17
#> 17  37  32  34  27  45  40  38  39  37  36
#> 18  43  41  36  32  28  41  33  36  33  44
#> 19  35  30  43  40  44  35  43  30  47  32
#> 20  37  36  43  35  36  40  33  40  18  51
#> 21  34  47  31  43  30  41  46  38  33  37
#> 22  44  34  31  39  36  29  35  49  40  34
#> 23  44  49  26  24  33  29  27  32  33  35
#> 24  42  32  45  37  31  29  32  30  39  32
#> 25  28  30  34  38  43  35  39  27  38  37
#> 26  30  44  48  37  32  41  27  33  26  42
#> 27  42  29  26  28  49  36  42  28  32  33
#> 28  26  34  39  37  24  35  37  29  30  40
#> 29  29  32  36  41  29  32  33  31  31  30
#> 30  32  34  27  36  29  30  26  41  45  36