Simulate microbiome perturbation data

To simulate microbiome longitudinal count data for perturbation experiments with user defined starting and ending time point, effect size and duration of perturbation

sim_ts_perturb(
  endtime,
  n_species,
  n_depth,
  Sigma = NULL,
  beta,
  t1,
  t2,
  time,
  species,
  duration
)

Arguments

endtime

the number of time points of the sample

n_species

number of species

n_depth

sequence depth

Sigma

correlation matrix for muti-nomal distribution

beta

effect of perturbation

t1

time of reaching the maximum of perturbation

t2

time of ending the maximum of perturbation

time

the start time for perturbation

species

perturbed species

duration

the duration of perturbation

Value

the simulated data

Examples

sim_ts_perturb(endtime=30, n_species=30, 1000, Sigma=NULL, beta=1,species=c(1,2,3),time=4, t1=8,t2=12,duration = 16)
#>     V1  V2  V3 V4 V5 V6 V7 V8 V9 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V20
#> 1   40  58  91 22 53 59 35 12 36  11  11  80  15  38  43  19  22  49  27  23
#> 2   41  41  62 27 36 58 27 25 33  23  37  58  24  48  29  27  27  38  25  34
#> 3   27  30  42 35 39 44 34 38 23  33  30  40  31  42  34  35  33  28  31  39
#> 4   38  35  46 32 34 35 44 25 45  25  36  24  32  36  35  28  44  30  26  34
#> 5   86  84  67 29 34 25 30 32 24  31  28  26  25  34  38  26  24  34  26  34
#> 6  225 164 197 22 17 11  9 19  7  19  15  25  12  22  23  18  12  11  12  14
#> 7  326 259 320  1  3  2  2  7  7   5   1   3   6   5   3   1   2   3   2   4
#> 8  309 329 344  2  1  0  0  0  2   1   1   0   1   0   2   0   0   0   0   0
#> 9  314 327 350  0  0  0  1  0  0   1   0   0   0   1   1   0   0   0   0   0
#> 10 312 325 358  1  0  0  1  0  0   0   0   0   0   0   0   0   1   0   0   0
#> 11 339 339 320  0  0  0  0  0  0   0   0   0   0   0   0   0   1   0   0   0
#> 12 312 369 317  0  0  0  0  0  0   0   0   0   0   0   1   0   0   0   0   0
#> 13 352 317 320  1  1  1  0  0  0   0   0   1   1   0   0   1   1   0   0   0
#> 14 290 343 331  0  0  3  2  3  0   1   3   2   3   3   0   2   0   0   1   1
#> 15 274 274 286  5  5 10  4  6  9   8   6   3   6   6   5   6   7   6   3   8
#> 16 141 156 133 22 23 26 15 19 22  12  21  26  23  19  20  18  15  25  25  29
#> 17  25  34  33 44 33 29 37 44 40  36  35  24  39  29  28  27  27  31  31  24
#> 18   4   6   4 36 32 34 35 38 31  41  44  23  34  25  40  36  33  34  48  39
#> 19   1   0   0 27 41 34 50 34 31  47  34  33  49  46  37  31  35  29  36  24
#> 20   0   0   0 50 39 36 36 42 33  37  40  35  36  35  34  36  37  35  36  35
#> 21   2   1   0 33 39 31 42 32 29  47  33  34  34  32  45  43  42  46  31  35
#> 22  10   8  11 34 37 31 41 30 33  38  35  40  36  30  44  24  29  44  45  38
#> 23  17  20  24 47 32 43 40 27 31  37  35  31  35  38  37  43  34  35  41  33
#> 24  16  29  20 32 22 40 29 28 37  34  37  30  39  28  40  44  26  41  36  32
#> 25  26  21  25 33 35 31 33 33 32  41  32  26  37  40  29  35  35  36  39  32
#> 26  33  31  28 40 37 36 29 32 34  38  36  48  27  31  32  32  41  23  22  35
#> 27  35  39  30 35 36 31 41 26 44  25  35  31  25  28  30  19  26  29  27  37
#> 28  33  25  35 48 39 33 42 18 31  28  35  37  31  29  30  33  33  42  34  30
#> 29  35  30  30 35 36 33 43 33 40  38  37  39  27  39  30  41  36  31  26  35
#> 30  30  28  36 31 26 29 35 31 35  35  38  39  30  37  40  39  26  33  28  42
#>    V21 V22 V23 V24 V25 V26 V27 V28 V29 V30
#> 1   12  36  22  22  33  30  26  30  14  31
#> 2   20  27  29  26  32  30  31  35  24  26
#> 3   19  39  36  39  30  39  27  30  24  29
#> 4   35  29  28  40  36  29  30  29  34  26
#> 5   20  23  26  22  35  26  24  30  26  31
#> 6   14  22  15  20  11  21  10  11   8  14
#> 7    2   5   4   3   3   2   6   3   2   8
#> 8    0   0   0   2   1   2   0   3   0   0
#> 9    1   1   1   0   0   1   0   0   0   1
#> 10   1   0   0   1   0   0   0   0   0   0
#> 11   1   0   0   0   0   0   0   0   0   0
#> 12   0   0   0   0   0   0   0   0   0   1
#> 13   1   0   1   0   0   0   0   1   0   1
#> 14   2   0   0   2   2   0   0   2   0   4
#> 15  10   6   8   9   6   7   7   2   6   2
#> 16  16  30  14  23  22  24  18  18  27  18
#> 17  43  27  33  31  44  38  29  36  32  37
#> 18  37  38  41  40  38  43  42  32  34  38
#> 19  31  32  35  38  36  44  42  49  36  38
#> 20  29  32  39  34  34  46  41  46  34  33
#> 21  42  39  26  25  38  43  39  42  34  41
#> 22  31  40  27  47  32  39  33  40  39  34
#> 23  37  32  32  27  38  32  36  30  24  32
#> 24  49  30  35  24  35  29  40  51  31  36
#> 25  27  43  30  42  42  31  24  40  44  26
#> 26  32  43  23  36  33  29  31  35  38  35
#> 27  34  27  43  29  51  45  35  32  32  43
#> 28  28  38  29  40  34  32  33  35  29  36
#> 29  39  33  16  30  29  29  28  37  28  37
#> 30  29  31  44  37  28  34  40  31  28  30