Simulate microbiome perturbation data

To simulate microbiome longitudinal count data for perturbation experiments with user defined starting and ending time point, effect size and duration of perturbation

sim_ts_perturb(
  endtime,
  n_species,
  n_depth,
  Sigma = NULL,
  beta,
  t1,
  t2,
  time,
  species,
  duration
)

Arguments

endtime

the number of time points of the sample

n_species

number of species

n_depth

sequence depth

Sigma

correlation matrix for muti-nomal distribution

beta

effect of perturbation

t1

time of reaching the maximum of perturbation

t2

time of ending the maximum of perturbation

time

the start time for perturbation

species

perturbed species

duration

the duration of perturbation

Value

the simulated data

Examples

sim_ts_perturb(endtime=30, n_species=30, 1000, Sigma=NULL, beta=1,species=c(1,2,3),time=4, t1=8,t2=12,duration = 16)
#>     V1  V2  V3 V4 V5 V6 V7 V8 V9 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V20
#> 1   17  27  51 27 43 12 66 32 25  28  51   6  51  21  30  14  31   7  14  19
#> 2   27  30  42 34 40 20 55 38 29  37  37  23  42  19  32  18  33  28  28  26
#> 3   35  32  37 24 32 32 46 32 41  19  34  21  41  24  25  30  31  29  38  28
#> 4   31  29  40 34 39 31 47 39 25  20  36  31  32  29  25  26  37  44  37  30
#> 5   68  84  99 35 31 35 30 25 21  40  24  22  26  22  29  32  31  26  20  33
#> 6  195 180 205 18 18 22 28 15 13  15  11  16  10  12  16  15  20  10  14  14
#> 7  284 330 290  6  3  3  4  4  3   4   4   3   2   1   6   4   2   4   2   1
#> 8  325 328 324  0  1  1  1  2  0   1   0   2   2   0   1   2   0   2   1   0
#> 9  340 325 326  0  0  0  0  2  0   0   0   2   0   0   1   1   0   0   0   1
#> 10 354 330 309  0  0  0  0  0  1   0   0   0   1   0   1   0   0   0   1   1
#> 11 332 336 328  0  0  0  0  0  1   0   0   0   0   0   0   0   0   0   0   1
#> 12 351 310 335  0  0  0  1  0  0   1   0   0   0   1   0   0   0   0   0   0
#> 13 303 357 330  1  0  0  2  0  0   0   0   1   0   1   0   0   0   0   0   0
#> 14 323 309 333  3  1  2  0  2  3   0   1   0   0   2   0   2   3   0   2   1
#> 15 266 225 314  6  7  7  9  6  9   9   7   3  11   7  13   8   1   9   6   5
#> 16 138 134 139 29 20 19 12 23 12  27  21  20  25  14  26  16  18  25  24  24
#> 17  28  29  25 27 28 34 28 22 31  36  32  39  23  39  34  24  31  35  36  41
#> 18   3   3   4 39 39 31 43 38 29  43  29  46  29  50  42  37  26  40  34  38
#> 19   1   3   1 34 44 43 36 28 38  52  50  41  32  26  35  33  23  40  28  34
#> 20   0   0   0 33 35 36 41 31 31  47  45  34  40  48  42  33  36  45  38  48
#> 21   2   2   3 43 31 34 31 34 39  38  38  40  42  31  46  27  49  31  34  34
#> 22  17   4   7 31 38 46 41 41 45  27  39  35  36  29  31  32  32  25  43  34
#> 23  18  16  22 43 40 32 37 41 22  33  31  28  34  43  37  29  27  37  42  25
#> 24  22  28  19 44 30 34 44 35 45  29  36  31  22  38  27  27  39  41  33  35
#> 25  20  29  23 30 38 29 33 34 39  37  50  31  34  35  26  28  21  33  33  27
#> 26  34  39  30 34 38 30 30 39 40  26  32  34  36  37  32  34  33  25  35  30
#> 27  32  42  45 27 26 36 32 39 43  38  27  29  39  33  28  37  34  27  37  40
#> 28  32  37  46 33 34 36 38 41 35  35  40  32  29  33  39  37  24  39  30  31
#> 29  39  32  25 30 37 25 36 37 33  37  29  33  38  30  31  33  38  46  33  36
#> 30  37  29  37 29 29 30 26 30 28  29  23  43  39  30  31  38  30  46  35  32
#>    V21 V22 V23 V24 V25 V26 V27 V28 V29 V30
#> 1   20  72  16  50  46  52  35  24  81  32
#> 2   26  54  19  40  33  53  21  33  48  35
#> 3   38  46  35  45  33  38  48  31  26  29
#> 4   33  40  27  35  39  25  40  33  37  29
#> 5   25  34  21  27  31  32  29  19  24  25
#> 6   11  20  16  18  12  27  12  11  16  10
#> 7    3   1   4   2   5   4   6   5   7   3
#> 8    0   0   1   1   1   0   0   3   0   1
#> 9    0   0   0   0   1   1   0   0   0   0
#> 10   0   0   0   0   0   0   1   1   0   0
#> 11   0   0   0   1   1   0   0   0   0   0
#> 12   0   0   0   0   0   0   1   0   0   0
#> 13   0   0   1   0   0   1   0   1   1   1
#> 14   0   2   2   2   2   1   1   0   2   1
#> 15   5   9   7   5  11   8   6  10   5   6
#> 16  29  20  24  22  21  21  22  24  21  30
#> 17  44  33  34  28  39  32  46  49  33  40
#> 18  37  44  30  31  45  41  26  30  30  43
#> 19  39  44  41  31  40  34  37  27  45  40
#> 20  28  34  38  32  30  34  39  37  32  33
#> 21  38  40  34  33  44  38  40  35  31  38
#> 22  51  31  41  39  34  39  28  34  38  32
#> 23  38  38  30  39  22  35  37  44  39  41
#> 24  28  35  36  44  28  23  43  37  38  29
#> 25  40  37  38  32  34  31  45  47  37  29
#> 26  25  28  38  34  43  28  40  33  32  31
#> 27  27  37  29  29  30  24  28  38  36  31
#> 28  27  33  33  27  22  41  28  32  34  22
#> 29  26  41  30  35  32  33  30  33  28  34
#> 30  23  42  28  43  38  28  32  33  43  39